move dateArr calculation to week-view
